Updating configuration files
